backened almost over

+++ backend/controller/goalController.js
@@ -1,11 +1,12 @@
-const Goal=require('../model/goalModel')
-const asyncHandler=require('express-async-handler')
+const Goal = require('../model/goalModel')
+const User = require('../model/userModel')
+const asyncHandler = require('express-async-handler')
 
 
-const getGoal = asyncHandler(async(req, res) => {
-    const goal=await Goal.find()
+const getGoal = asyncHandler(async (req, res) => {
+    const goal = await Goal.find({ user: req.user.id })
 
-    if(!goal){
+    if (!goal) {
         res.status(400)
         throw new Error('Goal not found')
     }
@@ -13,44 +14,71 @@ const getGoal = asyncHandler(async(req, res) => {
     res.status(200).json(goal)
 })
 
-const setGoal =  asyncHandler(async(req, res) => {
-    console.log(req.body)
+const setGoal = asyncHandler(async (req, res) => {
+    // console.log(req.body)
     if (!req.body.text) {
         res.status(400)
         throw new Error('add some data')
     }
-        
-    const goal=await Goal.create({
-        text:req.body.text
+
+    const goal = await Goal.create({
+        text: req.body.text,
+        user: req.user.id
     })
 
     res.status(200).json(goal)
 })
 
-const updateGoal =  asyncHandler(async(req, res) => {
+const updateGoal = asyncHandler(async (req, res) => {
     const data = req.params.id
-    const goal=await Goal.findById(req.params.id)
+    const goal = await Goal.findById(req.params.id)
 
-    if(!goal){
+    if (!goal) {
         res.status(400)
         throw new Error('Goal not found')
     }
 
-    const updatedGoal=await Goal.findByIdAndUpdate(req.params.id,req.body,{new:true})
+    const user = await User.findById(req.user.id)
+
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    if (goal.user.toString() != user.id) {
+        res.status(401)
+        throw new Error('User not Authorized  ')
+    }
+
+    const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, { new: true })
 
     res.status(200).json(updatedGoal)
 })
 
-const deleteGoal =  asyncHandler(async(req, res) => {
-    
-    const goal=Goal.findById(req.params.id)
+const deleteGoal = asyncHandler(async (req, res) => {
 
-    if(!goal){
+    const goal = await Goal.findById(req.params.id)
+
+    if (!goal) {
         res.status(400)
         throw new Error('Goal not found')
     }
+
+    const user = await User.findById(req.user.id)
+
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+
+    if (goal.user.toString() != user.id) {
+        res.status(401)
+        throw new Error('User not Authorized  ')
+    }
+
     await Goal.findByIdAndDelete(req.params.id)
-    res.status(200).json({id:req.params.id})
+    res.status(200).json({ id: req.params.id })
 })
 
 module.exports = {
